Collision Work
Each hit takes 10 points off of the life. I kill cannonball sprites on impact. The big bug now is that I am trying to update the sprite image based off of its life and the motion is glitching and one of my ships is not taking damage and I do not know why